requirements.txt
Django
django-allauth
django-bootstrap-form
django-cors-headers
django-rest-auth
djangorestframework
django-bootstrap-v5
influxdb-client
pycryptodome
requests
requests-oauthlib
requests-unixsocket
reportlab
wheel
Weasyprint
fcm-django
drf-yasg

